Model class is now storing raw model data only
The vertex buffer responsible for visual output now is created as well as stored separately.
This should make things easier to understand and untangle a few knots.
An issue appeared with the resizing of the raw model. This is originating from the new creation process and the usage of matrices. The resizing process needs to be rebuilt